アイテムを作成したユーザーのみが更新、削除できるようにする

+++ app/api/item/delete/[id]/route.js
@@ -5,8 +5,12 @@ import { NextResponse } from "next/server";
 export async function DELETE(request, context) {
   try {
     await connectDB();
-    await ItemModel.deleteOne({ _id: context.params.id });
-    return NextResponse.json({ message: "アイテム削除成功" });
+    const singleItem = await ItemModel.findById(context.params.id);
+    if (singleItem.email === reqBody.email) {
+      await ItemModel.deleteOne({ _id: context.params.id });
+      return NextResponse.json({ message: "アイテム削除成功" });
+    }
+    return NextResponse.json({ message: "他の人が作成したアイテムです" });
   } catch (error) {
     return NextResponse.json({ message: "アイテム削除失敗" });
   }
